pyproject.toml
[project]
name = "podcast-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.115.11",
    "google-genai>=1.5.0",
    "kokoro>=0.8.4",
    "librosa>=0.11.0",
    "matplotlib==3.7.0",
    "moviepy==1.0.3",
    "pip>=25.0.1",
    "soundfile>=0.13.1",
    "uvicorn>=0.34.0",
]

[tool.uv.workspace]
members = ["kokoro_agent"]
